removing old unused applicant task list code
