--- src/main/steps/applicant/task-list/section-titles.ts
@@ -1,21 +1,0 @@
-export const applicant_en = {
-  applicantYourDetails: 'About you',
-  yourApplication: 'Your application',
-  applicationDetails: 'Application detail',
-  viewAllDocuments: 'Your documents',
-  ordersFromTheCourt: 'Orders from the court',
-  courtHearings: 'Your court hearings',
-  yourDocuments: 'Your documents',
-  theResponse: 'response',
-};
-
-export const applicant_cy = {
-  applicantYourDetails: 'Amdanoch chi',
-  yourApplication: 'Y cais',
-  applicationDetails: 'Application detail',
-  viewAllDocuments: 'Eich dogfennau',
-  ordersFromTheCourt: 'Gorchmynion gan y llys',
-  courtHearings: 'Eich gwrandawiadau llys',
-  yourDocuments: 'Eich dogfennau',
-  theResponse: 'ymateb',
-};
