--- src/main/steps/applicant/task-list/tasklist-items.ts
@@ -1,39 +1,0 @@
-export const applicant_tasklist_items_en = {
-  keep_your_details_private: 'Keep your details private',
-  confirm_or_edit_your_contact_details: 'Confirm or edit your contact details',
-  your_application: 'Application submitted (PDF)',
-  your_application_ca: 'Your application (PDF)',
-  your_allegations_of_harm: 'View allegations of harm and violence (PDF)',
-  respond_to_other_side_aoh_violence: "Respond to the other side's allegations of harm and violence",
-  your_application_witness_statement: 'Witness statement (PDF)',
-  response_to_your_application: 'The response to your application (PDF)',
-  check_other_side_aoh_and_violence: "Check the other side's allegations of harm and violence",
-  view_all_documents: 'View all documents',
-  view_all_orders_from_the_court: 'View all orders from the court',
-  support_you_need_during_your_case: 'Support you need during your case',
-  application_submitted: 'Application submitted (PDF)',
-  witness_statement: 'Witness statement (PDF)',
-  details_of_court_hearings: 'Check details of your court hearings',
-  orders: 'View all orders from the court',
-  upload_document: 'Upload documents',
-};
-
-export const applicant_tasklist_items_cy = {
-  keep_your_details_private: 'Cadw eich manylion yn breifat',
-  confirm_or_edit_your_contact_details: 'Cadarnhau neu olygu eich manylion cyswllt',
-  your_application: 'Cais wedi’i gyflwyno (PDF)',
-  your_application_ca: 'Eich cais (PDF)',
-  your_allegations_of_harm: 'Gweld honiadau o niwed a thrais (PDF)',
-  respond_to_other_side_aoh_violence: "Respond to the other side's allegations of harm and violence",
-  your_application_witness_statement: 'Datganiad tyst (PDF)',
-  response_to_your_application: 'Gweld yr ymateb i’ch cais (PDF)',
-  check_other_side_aoh_and_violence: "Check the other side's allegations of harm and violence",
-  view_all_documents: 'Gweld yr holl ddogfennau',
-  view_all_orders_from_the_court: 'Gweld yr holl orchmynion gan y llys',
-  support_you_need_during_your_case: 'Cefnogaeth sydd ei hangen arnoch yn ystod eich achos',
-  application_submitted: 'Cais wedi’i gyflwyno (PDF)',
-  witness_statement: 'Datganiad tyst (PDF)',
-  details_of_court_hearings: 'Gwiriwch fanylion eich gwrandawiadau llys',
-  orders: 'Gweld yr holl orchmynion gan y llys',
-  upload_document: 'Llwytho dogfennau Dewisol',
-};
