--- src/main/steps/applicant/task-list/template.njk
@@ -1,35 +1,0 @@
-{% from "govuk/components/notification-banner/macro.njk" import govukNotificationBanner %}
-{% from "common/progress-bar.njk" import progressBar with context %}
-
-{% extends "common/tasklist/template.njk" %}
-{% block caseDetails %}
-{% endblock %}
-
-{% block columnFull %}
-<div class="govuk-grid-column-full">
-  <h1 class="govuk-heading-xl govuk-!-margin-bottom-4">
-    {{applicantName}}
-  </h1>
-  <p class="govuk-body">{{caseNumber}} {{userCase.id}}</p>
-</div>
-{% endblock %}
-
-{% block progressBar %}
-  {{ progressBar(stages) }}
-{% endblock %}
-
-{% block notificationBanner %}
-  {% include "../../common/notifyBanner.njk" %}
-{% endblock %}
-{% block oneThird %}
-      <h2 class="govuk-heading-m">{{iWantTo}}</h2>
-          <ul class="govuk-list">
-          {% for hyperlink in hyperlinks %}
-            <li>
-              <a href="{{hyperlink.link}}" class="govuk-link">{{hyperlink.label}}</a>
-              <div class="govuk-!-padding-bottom-2"></div>
-            </li>
-          {% endfor %}
-          </ul class="govuk-list">
-        {% include "../../common/contact-us.njk" %}
-{% endblock %}
