+++ src/main/steps/applicant/task-list/utils.test.ts
@@ -1,17 +1,11 @@
 import mockUserCase from '../../../../test/unit/utils/mockUserCase';
 import { CaseWithId } from '../../../app/case/case';
-import { SectionStatus, State, YesOrNo } from '../../../app/case/definition';
+import { State } from '../../../app/case/definition';
 
 import {
   getApplicantAllegationsOfHarmAndViolence,
-  getApplicantResponseToRequestForChildArrangements,
-  getApplicantViewAllOrdersFromTheCourtAllDocuments,
-  getConfirmOrEditYourContactDetails,
-  getKeepYourDetailsPrivateStatus,
-  getMiamStatus,
+  getApplicantPartyDetails,
   getSupportYourNeedsDetails,
-  getViewAllDocuments,
-  getYourApplication,
 } from './utils';
 
 const userCase: CaseWithId = {
@@ -22,107 +16,6 @@ const userCase: CaseWithId = {
 };
 
 describe('utils', () => {
-  describe('getKeepYourDetailsPrivateStatus', () => {
-    test.each([
-      {
-        data: {
-          ...mockUserCase,
-          detailsKnown: undefined,
-          startAlternative: undefined,
-        },
-        expected: SectionStatus.TO_DO,
-      },
-      {
-        data: {
-          ...mockUserCase,
-          detailsKnown: 'undefined',
-          startAlternative: 'undefined',
-        },
-        expected: SectionStatus.TO_DO,
-      },
-      {
-        data: {
-          ...mockUserCase,
-          detailsKnown: 'undefined',
-          startAlternative: undefined,
-        },
-        expected: SectionStatus.TO_DO,
-      },
-      // {
-      //   data: {
-      //     ...mockUserCase,
-      //     detailsKnown: 'undefined',
-      //     startAlternative: undefined,
-      //     confidentialDetails:'string',
-
-      //   },
-      //   expected: SectionStatus.IN_PROGRESS,
-      // },
-    ])('should return correct status %#', async ({ data, expected }) => {
-      expect(getKeepYourDetailsPrivateStatus({ ...userCase, ...data }, '123456')).toBe(expected);
-    });
-  });
-  describe('getConfirmOrEditYourContactDetails', () => {
-    test.each([
-      {
-        data: {
-          ...mockUserCase,
-          citizenUserFullName: undefined,
-          citizenUserDateOfBirth: undefined,
-        },
-        expected: SectionStatus.IN_PROGRESS,
-      },
-      {
-        data: {
-          ...mockUserCase,
-          citizenUserFullName: YesOrNo.NO,
-        },
-        expected: SectionStatus.IN_PROGRESS,
-      },
-      // {
-      //   data: {
-      //     ...mockUserCase,
-      //     citizenUserFullName:'Test',
-      //     citizenUserDateOfBirth: {
-      //       year: 'string',
-      //       month: 'string',
-      //       day: 'string',
-      //     },
-      //     citizenUserPlaceOfBirth: 'string',
-      //   },
-      //   expected: SectionStatus.COMPLETED,
-      // },
-      {
-        data: {
-          citizenUserFullName: 'Test',
-          citizenUserDateOfBirth: {
-            year: 'string',
-            month: 'string',
-            day: 'string',
-          },
-          citizenUserPlaceOfBirth: 'string',
-        },
-        expected: SectionStatus.IN_PROGRESS,
-      },
-    ])('should return correct status %#', async ({ data, expected }) => {
-      expect(getConfirmOrEditYourContactDetails({ ...userCase, ...data }, '123456')).toBe(expected);
-    });
-  });
-  describe('getYourApplication', () => {
-    test.each([
-      {
-        data: {
-          ...mockUserCase,
-          doYouConsent: YesOrNo.NO,
-          courtPermission: YesOrNo.NO,
-        },
-        expected: SectionStatus.DOWNLOAD,
-      },
-      { data: { ...mockUserCase, doYouConsent: YesOrNo.NO }, expected: SectionStatus.DOWNLOAD },
-    ])('should return correct status %#', async ({ expected }) => {
-      expect(getYourApplication()).toBe(expected);
-    });
-  });
   describe('getApplicantAllegationsOfHarmAndViolence', () => {
     test.each([
       {
@@ -141,145 +34,6 @@ describe('utils', () => {
       },
     ])('should return correct status %#', async ({ data, expected }) => {
       expect(getApplicantAllegationsOfHarmAndViolence({ ...userCase, ...data })).toBe(expected);
-    });
-  });
-  describe('getApplicantResponseToRequestForChildArrangements', () => {
-    test.each([
-      {
-        data: {
-          ...mockUserCase,
-          childrenKnownToLocalAuthority: '',
-        },
-        expected: false,
-      },
-      {
-        data: {
-          ...mockUserCase,
-          childrenKnownToLocalAuthority: 'YesOrNo.NO',
-        },
-        expected: true,
-      },
-    ])('should return correct status %#', async ({ data, expected }) => {
-      expect(getApplicantResponseToRequestForChildArrangements({ ...userCase, ...data })).toBe(expected);
-    });
-  });
-
-  describe('getApplicantViewAllOrdersFromTheCourtAllDocuments', () => {
-    test.each([
-      {
-        data: {
-          ...mockUserCase,
-          orderCollection: [],
-        },
-        expected: false,
-      },
-      {
-        data: {
-          ...mockUserCase,
-          orderCollection: [
-            {
-              id: '1234',
-              value: {
-                dateCreated: 'date',
-                orderType: 'type',
-                orderDocument: {
-                  document_url: 'string',
-                  document_filename: 'string',
-                  document_binary_url: 'string',
-                  document_hash: 'string',
-                },
-                orderDocumentWelsh: {
-                  document_url: 'string',
-                  document_filename: 'string',
-                  document_binary_url: 'string',
-                  document_hash: 'string',
-                },
-                otherDetails: {
-                  createdBy: 'string',
-                  orderCreatedDate: 'string',
-                  orderMadeDate: 'string',
-                  orderRecipients: 'string',
-                },
-              },
-            },
-          ],
-        },
-        expected: true,
-      },
-      {
-        data: {
-          ...mockUserCase,
-          orderCollection: [],
-        },
-        expected: false,
-      },
-      {
-        data: {
-          ...mockUserCase,
-          orderCollection: [],
-        },
-        expected: false,
-      },
-    ])('should return correct status %#', async ({ data, expected }) => {
-      expect(getApplicantViewAllOrdersFromTheCourtAllDocuments({ ...userCase, ...data })).toBe(expected);
-    });
-  });
-  describe('getYourSafetyStatus', () => {
-    test.each([
-      {
-        data: {
-          ...mockUserCase,
-        },
-        expected: SectionStatus.READY_TO_VIEW,
-      },
-    ])('should return correct status %#', async ({ expected }) => {
-      expect(getViewAllDocuments()).toBe(expected);
-    });
-  });
-  describe('getMiamStatus', () => {
-    test.each([
-      {
-        data: {
-          ...mockUserCase,
-          miamStart: 'Yes',
-          miamWillingness: 'Yes',
-        },
-        expected: 'COMPLETED',
-      },
-      {
-        data: {
-          ...mockUserCase,
-          miamStart: '',
-          miamWillingness: '',
-        },
-        expected: 'TO_DO',
-      },
-      {
-        data: {
-          ...mockUserCase,
-          miamStart: 'Yes',
-          miamWillingness: undefined,
-        },
-        expected: 'IN_PROGRESS',
-      },
-      {
-        data: {
-          ...mockUserCase,
-          miamStart: undefined,
-          miamWillingness: 'Yes',
-        },
-        expected: 'IN_PROGRESS',
-      },
-      {
-        data: {
-          ...mockUserCase,
-          miamStart: undefined,
-          miamWillingness: undefined,
-        },
-        expected: 'TO_DO',
-      },
-    ])('should return correct status %#', async ({ data, expected }) => {
-      expect(getMiamStatus(data)).toBe(expected);
     });
   });
 
@@ -317,5 +71,13 @@ describe('utils', () => {
     ])('should return correct status %#', async ({ data, expected }) => {
       expect(getSupportYourNeedsDetails(data)).toBe(expected);
     });
+  });
+  test('getApplicantPartyDetails', () => {
+    const data = {
+      id: '12',
+      state: State.CASE_SUBMITTED_NOT_PAID,
+    };
+
+    expect(getApplicantPartyDetails(userCase, data.id)).toStrictEqual(undefined);
   });
 });
