+++ src/main/steps/common/add-legal-representative/content.ts
@@ -1,7 +1,7 @@
 import { PartyType } from '../../../app/case/definition';
 import { TranslationFn } from '../../../app/controller/GetController';
 import { FormContent } from '../../../app/form/Form';
-import { getApplicant, getApplicantName } from '../../applicant/task-list/content';
+import { getApplicant, getApplicantName } from '../../applicant/task-list';
 import { getCasePartyType } from '../../prl-cases/dashboard/utils';
 import { getRespondent, getRespondentName } from '../../respondent/task-list/content';
